BUGFIX: AG_USER_INFO_ALL now matches the AI server, so existing user info is sent on reconnection. We also free old AI session data (bit tacky, but it's better than a mem leak for now).

+++ Ebenezer/EbenezerDlg.cpp
@@ -1496,7 +1496,7 @@ BOOL CEbenezerDlg::LoadBlockNameList()
 void CEbenezerDlg::SendAllUserInfo()
 {
 	Packet result(AG_USER_INFO_ALL);
-	uint16 count = 0;
+	uint8 count = 0;
 	result << count; // placeholder for user count
 	const int tot = 20;
